# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for the Bug Bounty Toolkit application.",
        "title": "Toolkit API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v1.0.0"
    },
    "host": "localhost:8778",
    "basePath": "/api",
    "paths": {
...
        "/platforms": {
            "get": {
                "description": "Retrieves a list of all platforms, ordered by name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platforms"
                ],
                "summary": "List all platforms",
                "responses": {
                    "200": {
                        "description": "A list of platforms",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Platform"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new platform to the system. The platform name must be unique.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platforms"
                ],
                "summary": "Create a new platform",
                "parameters": [
                    {
                        "description": "Platform object that needs to be added",
                        "name": "platform",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Platform"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Platform created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Platform"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or missing name",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Platform with this name already exists",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/relationships/create": {
            "post": {
                "description": "(Not Implemented Yet) Manually links a discovered web page to a discovered API endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mapping"
                ],
                "summary": "Create a page-to-API relationship",
                "parameters": [
                    {
                        "description": "Request body containing web_page_id and api_endpoint_id",
                        "name": "relationship_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "501": {
                        "description": "Not Implemented Yet",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error message describing the issue"
                }
            }
        },
        "models.Platform": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "description": "Unique identifier for the platform.",
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "name": {
                    "description": "Name of the platform.",
                    "type": "string",
                    "example": "HackerOne"
                }
            }
        },
...
    }
}